fixed error for tsk 10

+++ 0x00-MySQL_Advanced/10-div.sql
@@ -3,9 +3,9 @@
 -- returns 0 if the second number is equal to 0.
 DELIMITER //
 CREATE FUNCTION SafeDiv(a INT, b INT)
-RETURNS DECIMAL(10,2)
+RETURNS FLOAT
 BEGIN
-    DECLARE result DECIMAL(10,2);
+    DECLARE result FLOAT;
     IF b = 0 THEN
         SET result = 0;
     ELSE
